.github/scripts/requirements.txt
playwright>=1.40.0
Pillow>=10.0.0
